backend/requirements.txt
aerich==0.8.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.5.0
appdirs==1.4.4
async-timeout==5.0.1
asyncclick==8.1.7.2
asyncpg==0.29.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dictdiffer==0.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-cli==0.0.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.3
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iso8601==2.1.0
jedi==0.19.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
parso==0.8.4
pip==24.3.1
prompt_toolkit==3.0.48
ptpython==3.0.29
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypika-tortoise==0.2.1
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
setuptools==75.3.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.2
tomlkit==0.13.2
tortoise-orm==0.21.7
typer==0.12.5
typer-slim==0.12.5
typing_extensions==4.12.2
uvicorn==0.32.1
watchfiles==0.23.0
wcwidth==0.2.13
websockets==13.0
wheel==0.45.1
